fix gcc warnings

+++ Repo.cpp
@@ -7,7 +7,7 @@
 BEGIN_INANITY_OIL
 
 const char Repo::protocolMagic[14] = { 'I', 'N', 'A', 'N', 'I', 'T', 'Y', 'O', 'I', 'L', 'R', 'E', 'P', 'O' };
-const int Repo::protocolVersion = 1;
+const size_t Repo::protocolVersion = 1;
 
 const int Repo::serverRepoAppVersion = 0x414C4941; // "OILA" in little-endian
 const int Repo::clientRepoAppVersion = 0x314C494F; // "OIL1" in little-endian
